move dkg bundle creation to Decrypter
